Create Homepage and demopage, add links to all routes
