+++ client/src/screens/index.tsx
@@ -1,3 +1,4 @@
 export {default as HomePage} from './Home';
 export {default as ComposePage} from './Compose';
 export {default as AuthPage} from './Auth';
+export {default as DemoPage} from './Demo';
